# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Engify AI API
  description: |
    Enterprise-grade API for the Engify AI prompt engineering platform.

    ## Features
    - 🔐 Secure authentication with NextAuth.js
    - 🎯 Rate limiting for fair usage
    - 📊 AI-powered chat with RAG (Retrieval Augmented Generation)
    - 🎨 100+ expert-crafted prompts
    - 🧩 15+ proven prompt patterns
    - 🔧 Multi-provider AI execution (OpenAI, Claude, Gemini, Groq, Replicate)

    ## Authentication
    Most endpoints require authentication via session cookies (NextAuth.js).
    Admin endpoints require elevated privileges.

    ## Rate Limits
    - **Anonymous**: 20 requests/hour, 50 requests/day
    - **Authenticated**: 100 requests/hour, 500 requests/day
    - **Pro**: 1000 requests/hour, 5000 requests/day

    Rate limit headers are included in responses:
    - `X-RateLimit-Limit`: Maximum requests allowed
    - `X-RateLimit-Remaining`: Requests remaining in current window
    - `X-RateLimit-Reset`: Time when the rate limit resets (ISO 8601)

  version: 2.0.0
  contact:
    name: Engify AI Support
    email: donlaur@engify.ai
    url: https://engify.ai
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://engify.ai/api
    description: Production server
  - url: http://localhost:3000/api
    description: Development server
...
paths:
  /chat:
    post:
      tags:
        - Chat
      summary: Send chat message
      description: |
        Send a message to the AI assistant. Supports RAG (Retrieval Augmented Generation)
        to provide contextually relevant responses from the knowledge base.

        **Rate Limits:**
        - Anonymous: 20/hour
        - Authenticated: 100/hour

        **Features:**
        - Automatic context retrieval from prompt library
        - Fallback to knowledge-based responses
        - Sanitized inputs and outputs
      operationId: sendChatMessage
      security:
        - sessionAuth: []
        - anonymous: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - messages
              properties:
                messages:
                  type: array
                  description: Chat message history
                  items:
                    type: object
                    required:
                      - role
                      - content
                    properties:
                      role:
                        type: string
                        enum: [user, assistant, system]
                        description: Message role
                      content:
                        type: string
                        description: Message content
                        maxLength: 10000
                useRAG:
                  type: boolean
                  default: true
                  description: Enable RAG for contextual responses
            examples:
              basicChat:
                summary: Basic chat message
                value:
                  messages:
                    - role: user
                      content: What is chain of thought prompting?
                  useRAG: true
              conversationWithHistory:
                summary: Chat with history
                value:
                  messages:
                    - role: user
                      content: Tell me about prompt patterns
                    - role: assistant
                      content: Prompt patterns are proven techniques...
                    - role: user
                      content: Which one is best for complex reasoning?
                  useRAG: true
      responses:
        '200':
          description: Successful response
          headers:
            X-RateLimit-Limit:
              schema:
                type: integer
              description: Maximum requests allowed
            X-RateLimit-Remaining:
              schema:
                type: integer
              description: Requests remaining
            X-RateLimit-Reset:
              schema:
                type: string
                format: date-time
              description: Rate limit reset time
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: AI response
                  sources:
                    type: array
                    description: Knowledge base sources used
                    items:
                      type: object
                      properties:
                        title:
                          type: string
                        score:
                          type: number
                  usedRAG:
                    type: boolean
                    description: Whether RAG was used
              examples:
                successWithRAG:
                  summary: Success with RAG
                  value:
                    message: "Chain of Thought (CoT) is one of our most powerful patterns..."
                    sources:
                      - title: "Chain of Thought Pattern"
                        score: 0.95
                      - title: "Advanced Reasoning Techniques"
                        score: 0.87
                    usedRAG: true
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate limit exceeded
          headers:
            Retry-After:
              schema:
                type: integer
              description: Seconds to wait before retrying
            X-RateLimit-Reset:
              schema:
                type: string
                format: date-time
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: free_trial_limit_reached
                  message:
                    type: string
                    example: You've reached the limit of your free trial. Sign in to continue.
                  resetAt:
                    type: string
                    format: date-time
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: |
        NextAuth.js session cookie. Automatically set after successful login.
        No need to manually include in requests when using a browser.
    anonymous:
      type: http
      scheme: bearer
      description: Anonymous access (no authentication required, but rate limited)
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error code or type
        message:
          type: string
          description: Human-readable error message
        details:
          type: array
          description: Additional error details
          items:
            type: object
      required:
        - error
      example:
        error: "validation_error"
        message: "Invalid input provided"
        details:
          - field: "email"
            issue: "Invalid email format"
...